+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added way to type a field as being generically JSON seralisable by using
+  the `JSONMixin` class.
 
 ## [0.3.1] - 2021-11-12
 ### Fixed